Re-merge IUiaWindow+IConsoleWindow, move back into Interactivity
