+++ src/host/selection.hpp
@@ -21,7 +21,7 @@ Revision History:
 #include "input.h"
 
 #include "..\interactivity\inc\IAccessibilityNotifier.hpp"
-#include "..\types\IConsoleWindow.hpp"
+#include "..\interactivity\inc\IConsoleWindow.hpp"
 
 class Selection
 {
